Move vampiric strength into its own component
